pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edmonton-growth"
version = "0.1.0"
description = "Edmonton Urban Growth Intelligence Dashboard"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyyaml>=6.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pyproj>=3.6.0",
    "fiona>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]
